All tutorials
These are the class where all selenium tests are, a quick example, and the generic methods used in the project.
allTestCases is where all the tests are.
BasicActions was just a quick practice/example file.
GenericMethods is, as its name suggests, where all generic methods that are frequently used are located.